refactor(core): Split local_db by table: source & template
This also changes the behavior of `boil source list` to use on the
database rather than the local config file. This is because after a
successful `boil source add ...`, we do update the DB but the local
config is left untouched.

This was confusing: adding a source, and not seeing it. That's fixed,
and the display is in line with other commands (table based).